Polish `FlowServices`
- Simplify service registration wherever possible
- Move FlowServicesProvider to FlowServices
